refactor: 도메인 패키지 의존방향 정리
