+++ src/main/java/back/ecommerce/auth/dto/response/Token.kt
@@ -1,3 +1,3 @@
-package back.ecommerce.auth.token
+package back.ecommerce.auth.dto.response
 
 data class Token(val value: String, val expireTime: Int, val type: String)
