+++ src/test/java/back/ecommerce/auth/interceptor/AdminAuthorizationInterceptorTest.java
@@ -5,6 +5,7 @@ import static org.mockito.BDDMockito.*;
 
 import java.util.Optional;
 
+import back.ecommerce.api.auth.interceptor.AdminAuthorizationInterceptor;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
